backend/requirements.txt
flask
flask-socketio
python-socketio
eventlet>=0.24.1
opencv-python
numpy
ultralytics
python-dotenv
groq
gunicorn